+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 * [UIRS-13](https://issues.folio.org/browse/UIRS-13) Update stripes to v6.0.0
 * [UIRS-19](https://issues.folio.org/browse/UIRS-19) Update stripes-cli to v2
 * [UIRS-10](https://issues.folio.org/browse/UIRS-10) Add personal data disclosure form
+* [UIRS-20](https://issues.folio.org/browse/UIRS-20) Rename accession and retrieval configurations